4.23 done

+++ part 4/blogilista/controllers/blogs.js
@@ -1,26 +1,43 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const { userExtractor } = require("../utils/middleware");
+const jwt = require("jsonwebtoken");
+
+//GET
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response, next) => {
+//POST
+
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const body = request.body;
 
-  if (!body.likes) body.likes = 0;
+  const { user } = request;
+
+  if (!user) {
+    return response.status(401).json({ error: "token invalid" });
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes,
+    user: user,
+    likes: body.likes ? body.likes : 0,
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
+
+//GET:id
 
 blogsRouter.get("/:id", async (request, response) => {
   const blog = await Blog.findById(request.params.id);
@@ -31,10 +48,26 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+//DELETE
+
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const { user } = request;
+
+  if (!user) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (blogToDelete.user._id.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: "Unauthorized" });
+  }
 });
+
+//PUT
 
 blogsRouter.put("/:id", async (request, response) => {
   const body = request.body;
@@ -53,27 +86,5 @@ blogsRouter.put("/:id", async (request, response) => {
   });
   response.json(updatedBlog.toJSON());
 });
-
-/*blogsRouter.get("/", (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs);
-  });
-});
-
-blogsRouter.post("/", (request, response, next) => {
-  const body = request.body;
-
-  const blog = new Blog({
-    content: body.content,
-  });
-  blog
-    .save()
-    .then((savedBlog) => {
-      response.json(savedBlog);
-    })
-    .catch((error) => next(error));
-});
-
-*/
 
 module.exports = blogsRouter;
